Add an implementation of the block cache on NVM
This commit adds support for a persistent memory cache based on libnvml[1].
The cache is configured using a new flag --block_cache_type=NVM and
--nvm_cache_path=/path/to/mountpoint/. The mointpoint is expected
to be a filesystem which supports DAX[2] mounted on a persistent memory device.

For testing purposes, the new cache implementation can also be run on
any old other mount point (eg /dev/shm or a normal file system) though
there is not likely to be a performance advantage.

The code changes contained are:

* introduce the new flags to select the cache implementation
* introduce the new cache implementation, which uses libvmem APIs to
  allocate/free memory
* change CFileReader to read data directly from disk into cache-allocated
  memory when reading uncompressed data. When reading compressed data,
  we read into a DRAM scratch buffer which is uncompressed directly into
  the cache-allocated memory. This avoids an extra memcpy on cache misses
  when the new cache implementation is configured, but doesn't add any
  overhead in the DRAM case ("cache-allocated" memory is normal memory).
* cfile-test and block_cache-test are made into value-parameterized tests
  and test with both the DRAM and NVM caches.

This patch was primarily contributed by Sarah Jelinek, with some tweaks
by Todd Lipcon.

[1] http://pmem.io
[2] https://lwn.net/Articles/588218/

Change-Id: Ie2705e2deb0a0be8feac2fb7f5ec81f635613dbc